+++ CHANGELOG.md
@@ -4,11 +4,17 @@
 
 🎉🎉 YAML typed manifest and archived manifest installation support 🎉🎉
 
+- Prompt for the new issue now has correct version in title in case of archived manifest is being downloaded/installed/...
+- **Unix**:
+    - Fix `--version` command under *nix
+    - Supress errors about environment variables, until fully implemented.
+        - 🎉 Manifest installation on Unix is now partially implemented
 - **CI**: Support validation of YML typed manifests
 - Cache the arm check to prevent multiple calls of `uname` on *nix
-- **Unix**: Fix `--version` command under *nix
 - **scoop-uninstall**: Warn user when he tries to uninstall dependency, which is still needed
-- **scoop-status**: Detect if applications, which were installed as dependency are still needed
+- **scoop-status**:
+    - Properly support yml typed manifests and dependencies
+    - Detect if applications, which were installed as dependency are still needed
 - **scoop-list**:
     - Fix detection of installed status for PowerShell 5
     - Show if application was installed as dependency